+++ Lines/SegmentLineRenderer.cs
@@ -1,47 +1,44 @@
-//
+﻿//
 // Copyright (c) 2017 Geri Borbás http://www.twitter.com/_eppz
 //
 //  Permission is hereby granted, free of charge, to any person obtaining a copy of this software and associated documentation files (the "Software"), to deal in the Software without restriction, including without limitation the rights to use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of the Software, and to permit persons to whom the Software is furnished to do so, subject to the following conditions:
 //  The above copyright notice and this permission notice shall be included in all copies or substantial portions of the Software.
 //  THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 //
+#if EPPZ_LINES
 using UnityEngine;
 using System.Collections;
-using UnityEngine.Serialization;
 
 
-namespace EPPZ.Geometry.Components
+namespace EPPZ.Geometry.Lines
 {
 
 
-	public class PolygonSource : MonoBehaviour
+	using EPPZ.Lines;
+	using Components;
+
+
+	public class SegmentLineRenderer : GeometryLineRenderer
 	{
 
 
-		public Transform[] pointTransforms;
-		public Polygon.WindingDirection windingDirection = Polygon.WindingDirection.Unknown;
-		[Range (-2,2)] public float offset = 0.0f;
-		public bool updateModel = false;
-
-		public Polygon polygon;
-
-
-		void Awake()
+		public Color lineColor;
+		public Color boundsColor;
+		private Segment segment;
+		
+		
+		void Start()
 		{
-			// Construct a polygon model from transforms (if not created by a root polygon already).
-			if (polygon == null) polygon = Polygon.PolygonWithSource(this);
-			if (offset != 0.0f) polygon = polygon.OffsetPolygon(offset);
+			// Model reference.
+			SegmentSource segmentSource_ = GetComponent<SegmentSource>();
+			segment = segmentSource_.segment;
 		}
-
-		void Update()
+		
+		protected override void OnDraw()
 		{
-			if (updateModel)
-			{
-				// Update polygon model with transforms, also update calculations.
-				polygon.UpdatePointPositionsWithSource(this);
-				windingDirection = polygon.windingDirection;
-				if (offset != 0.0f) polygon = polygon.OffsetPolygon(offset);
-			}
+			DrawRect(segment.bounds, boundsColor);
+			DrawSegment(segment, lineColor);
 		}
 	}
 }
+#endif